requirements.txt
streamlit
langchain
langgraph
langchain-groq
langchain-community
faiss-cpu
sentence-transformers
python-dotenv